Creates basic logic for RPS Game
The playerchoice is hardcoded to be rock as of this commit.
However, the computerSelection is randomized and there is parsing
to somewhat error check and also accounts for any permutation
of input for player, assuming it is a valid string.
If for some reason it isn't, a faulty gamestate is printed.